requirements.txt
Flask==3.1.1
python-docx==1.1.0
google-generativeai==0.8.3
python-dotenv==1.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0

